switching register logic to user service
